feat(effect): §C SUPPRESS — disable enchant/group/type (DISABLE_*)
Completes §C's remaining primitives: SUPPRESS temporarily disables a target
player's enchant / group / type for a span of ticks (AE's DISABLE_ENCHANT /
DISABLE_GROUP / DISABLE_TYPE).

Design — reuse the three cooldown scopes (no new bridge interner): an enchant's
cdScopeEnchant/cdScopeGroup/cdScopeType are already populated (baseKey / group /
type) and interned in the cooldownScopes namespace. So:

- SuppressEffect (scope, key, duration; default target VICTIM) emits Sink.suppress.
- The compile EraseStage rewrites a SUPPRESS effect's scope→kind int and key→its
  cooldownScopes interner id — the SAME id the suppressed abilities lower their
  scope to (the bridge invariant, pinned by EraseStageTest).
- DispatchSink.suppress writes the SuppressionStore keyed by CooldownStore.key(
  scopeKind, scopeId) — the in-memory, concurrent, lazy-TTL store (rekeyed int→long).
- Gate 5 of the ActivationPipeline checks the ability's three scopes against the
  store, mirroring gate 6's cooldown check exactly (O(1), hot-path-safe).

Wiring: one shared SuppressionStore threaded from the bootstrap into the pipeline
(gate-5 read) and both dispatchers (per-event sink write); cleared on quit via the
EngineStoreListener. DispatchSink gains a 6th ctor arg; the convenience ctors
default it so all existing call sites compile unchanged.

Tests: SuppressEffectTest, the gate's timed/scoped suppression cases, the
erase-stage bridge invariant, and the sink write-through. ./gradlew build green.

With this, all four subsystem-coupled §C primitives are done (REMOVE_SOULS,
SET_VAR/INVERT_VAR, IGNORE_ARMOR, SUPPRESS). The combat-flags cluster
(KNOCKBACK_CONTROL / GUARD / KEEP_ON_DEATH) remains as a separate subsystem.

+++ se/engine/src/engine/stores/SuppressionStore.java
@@ -23,16 +23,18 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public final class SuppressionStore {
 
-    private final Map<UUID, Map<Integer, Long>> expiryByPlayer = new ConcurrentHashMap<>();
+    private final Map<UUID, Map<Long, Long>> expiryByPlayer = new ConcurrentHashMap<>();
 
     /**
-     * Suppress interned {@code id} for {@code durationTicks}, expiring at
-     * {@code nowTicks + durationTicks}. A non-positive duration is a no-op (no
-     * suppression). Re-suppressing an already-suppressed id only ever EXTENDS it: the
-     * expiry is set to the later of the existing and the new one, so a fresh short
+     * Suppress packed scope key {@code id} for {@code durationTicks}, expiring at
+     * {@code nowTicks + durationTicks}. The key is a {@link CooldownStore#key(int, int)}-packed
+     * (scopeKind, scopeId) pair so the three DISABLE scopes (enchant/group/type) never collide —
+     * shared with the gate's cooldown-scope namespace, so a {@code SUPPRESS} keys the same id the
+     * suppressed abilities lower their scope to. A non-positive duration is a no-op. Re-suppressing
+     * an already-suppressed key only ever EXTENDS it (the later expiry wins), so a fresh short
      * suppression never cuts an in-flight longer one short.
      */
-    public void suppress(UUID player, int id, long nowTicks, int durationTicks) {
+    public void suppress(UUID player, long id, long nowTicks, int durationTicks) {
         if (durationTicks <= 0) {
             return;
         }
@@ -47,8 +49,8 @@ public final class SuppressionStore {
      *     reported as not suppressed; the expiry tick itself counts as elapsed (the
      *     suppression covers {@code [start, expiry)}).
      */
-    public boolean isSuppressed(UUID player, int id, long nowTicks) {
-        Map<Integer, Long> ids = expiryByPlayer.get(player);
+    public boolean isSuppressed(UUID player, long id, long nowTicks) {
+        Map<Long, Long> ids = expiryByPlayer.get(player);
         if (ids == null) {
             return false;
         }
